workers/extract/pyproject.toml
[project]
name = "extract-worker"
version = "0.1.0"
description = "Long-running EC2 worker for document extraction via Step Functions Activity Task"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.76",
    "docling>=2.82.0",
    "python-magic>=0.4.27",
    "shared",
]
